written ocaml code to convert array to list then print it to screen
Used convert array to list by flattening the array out and making it a
single dimension list.

The list is traversed through by another function to convert the integers
to string and appends them to a list.  This is then con-catted
together separated by spaces into a single string.

The single string that is produced is the printed to the screen.