Basic rule checking on papers and fill in missing takes place
Does basic rules checking on papers and attempts to fill in
blanks. Still have some teething problems, like trying to add 161.1xx
as a paper.  It should not even give them the option to add this, but
just list this as a missing requirement.

Not all papers prerequisites are checked as some have rules which I
didn't take into account when designing this program, such as 2 200
level papers from BBS of BSc, or 2 of 159.101, 134.555, 123.456,
233.777.

I also don't do corequisites or restrictions, due to the fact I don't
have very much time.